requirements.txt
azure_storage==0.37.0
beautifulsoup4==4.12.3
Flask==3.0.2
Flask_Cors==4.0.0
Flask_PyMongo==2.3.0
nltk==3.8.1
pymongo==4.6.2
python-dotenv==1.0.1
Requests==2.31.0
selenium==4.19.0
Unidecode==1.3.8
webdriver_manager==4.0.1
azure-ai-textanalytics==5.3.0
gunicorn
openai